+++ src/modules/rt_events/CMakeLists.txt
@@ -15,6 +15,6 @@ forte_add_module(RT_Events OFF "Real-Time Event Funktion blocks")
 #############################################################################
 # RT_Events FB
 #############################################################################
-forte_add_sourcefile_hcpp(RT_E_CYCLE  RT_E_DEMUX       RT_E_F_TRIG  RT_E_PERMIT  RT_E_R_TRIG  RT_E_SPLIT)
-forte_add_sourcefile_hcpp(RT_E_DELAY  RT_E_EC_COUPLER  RT_E_MERGE   RT_E_REND    RT_E_SELECT  RT_E_SWITCH RT_E_TRAIN)
-forte_add_sourcefile_hcpp(RT_Bridge_fbt)
+forte_add_sourcefile_hcpp(RT_E_CYCLE  RT_E_DEMUX_fbt       RT_E_F_TRIG  RT_E_PERMIT  RT_E_R_TRIG  RT_E_SPLIT_fbt)
+forte_add_sourcefile_hcpp(RT_E_DELAY  RT_E_EC_COUPLER  RT_E_MERGE   RT_E_REND    RT_E_SELECT  RT_E_SWITCH_fbt RT_E_TRAIN)
+forte_add_sourcefile_hcpp(RT_Bridge_fbt_fbt)
