+++ src/modules/rt_events/RT_E_SWITCH_fbt.cpp
@@ -10,9 +10,9 @@
  *   Alois Zoitl, Gerhard Ebenhofer, Ingo Hegny, Monika Wenger,
  *    - initial API and implementation and/or initial documentation
  *******************************************************************************/
-#include "RT_E_SWITCH.h"
+#include "RT_E_SWITCH_fbt.h"
 #ifdef FORTE_ENABLE_GENERATED_SOURCE_CPP
-#include "RT_E_SWITCH_gen.cpp"
+#include "RT_E_SWITCH_fbt_gen.cpp"
 #endif
 
 #include "criticalregion.h"
